+++ clang/docs/ReleaseNotes.rst
@@ -92,6 +92,11 @@ Windows Support
 C Language Changes in Clang
 ---------------------------
 
+C2x Feature Support
+-------------------
+
+- Implemented `WG14 N2674 The noreturn attribute <http://www.open-std.org/jtc1/sc22/wg14/www/docs/n2764.pdf>`_.
+
 C++ Language Changes in Clang
 -----------------------------